backend/requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# LangGraph and LangChain (updated compatible versions)
langgraph==0.2.55
langchain-openai==0.2.10
langchain-core==0.3.28

# OpenAI
openai==1.58.0
httpx==0.27.2

# Data Validation
pydantic==2.10.2
pydantic-settings==2.6.1

# PDF and Image Processing
PyMuPDF==1.24.12
Pillow==11.0.0
opencv-python-headless==4.10.0.84

# Excel Export
openpyxl==3.1.5

# Redis (Optional)
redis==5.2.0

# Utilities
aiofiles==24.1.0
python-dotenv==1.0.1

# Fuzzy Matching
rapidfuzz==3.9.7

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
